Remove the clientType static field. Use class hierarchy for this.
Should allow running self-tests in parallel, and clean-up the code a bit.
