+++ cli-artemis-jms/src/main/java/com/redhat/mqe/acc/AccClientOptionManager.java
@@ -131,4 +131,9 @@ class AccClientOptionManager extends ClientOptionManager {
         }
         super.createConnectionOptions(clientOptions);
     }
+
+    @Override
+    protected String getUrlProtocol() {
+        return "tcp";
+    }
 }
