+++ cli-qpid-jms/src/main/java/com/redhat/mqe/jms/AacConnectionManagerFactory.java
@@ -23,7 +23,7 @@ import com.redhat.mqe.lib.ClientOptions;
 import com.redhat.mqe.lib.ConnectionManagerFactory;
 
 public class AacConnectionManagerFactory extends ConnectionManagerFactory {
-    AacConnectionManagerFactory() {
+    public AacConnectionManagerFactory() {
     }
 
     public AacConnectionManager make(ClientOptions clientOptions, String brokerUri) {
