+++ jmslib/src/main/java/com/redhat/mqe/lib/ConnectionManager.java
@@ -42,14 +42,9 @@ import java.util.Properties;
  */
 abstract public class ConnectionManager {
     protected ConnectionFactory factory;
-    protected InitialContext initialContext;
-    private Properties properties;
-    protected Destination destination;
     protected Connection connection;
-    protected String amqConnectionFactoryJNDI = "amqFactory";
-    protected String connectionFactory = "amqFactory"; // default amqp://localhost:5672
+    protected Destination destination;
     protected String password;
-    protected String queueOrTopic = "amqQueue";
     protected String username;
     protected static final String QUEUE_OBJECT = "javax.jms.Queue";
     protected static final String TOPIC_OBJECT = "javax.jms.Topic";
@@ -60,7 +55,6 @@ abstract public class ConnectionManager {
      * private static final String ARTEMIS_INITIAL_CONTEXT = "org.apache.activemq.artemis.jndi.ActiveMQInitialContextFactory";
      */
 
-    private static final String EXTERNAL_JNDI_PROPERTY = "aoc7.jndi";
     private Logger LOG = LoggerFactory.getLogger(ConnectionManager.class.getName());
 
     public ConnectionManager() {
@@ -75,19 +69,9 @@ abstract public class ConnectionManager {
         return destination;
     }
 
-    /**
-     * @param connectionFactory - often referred to as broker url
-     */
-    void setConnectionFactory(String connectionFactory) {
-        this.connectionFactory = connectionFactory;
-    }
+    protected abstract Queue createQueue(String queueName);
 
-    /**
-     * @param queueOrTopic - destination can be either queue or topic
-     */
-    void setDestinationName(String queueOrTopic) {
-        this.queueOrTopic = queueOrTopic;
-    }
+    protected abstract Topic createTopic(String topicName);
 
     /**
      * MessagingExceptionListener is created for each connection made.
@@ -98,95 +82,6 @@ abstract public class ConnectionManager {
             LOG.error("ExceptionListener error detected! \n{}\n{}", e.getMessage(), e.getCause());
             e.printStackTrace();
             System.exit(1);
-        }
-    }
-
-    /**
-     * Creates a destination for a specified name.
-     *
-     * @param destination for which destination is to be created.
-     * @return created Destination object
-     */
-    Destination createDestination(String destination) {
-        return (Destination) createJMSProviderObject("destination", destination);
-    }
-
-    /**
-     * Create queue object
-     *
-     * @param queueName name of the queue to be created
-     * @return created Queue object
-     */
-    protected Queue createQueue(String queueName) {
-        return (Queue) createJMSProviderObject("queue", queueName);
-    }
-
-    /**
-     * Create topic object
-     *
-     * @param topicName name of the topic to be created
-     * @return created Topic object
-     */
-    protected Topic createTopic(String topicName) {
-        return (Topic) createJMSProviderObject("topic", topicName);
-    }
-
-    /**
-     * Creates an object using qpid/amq initial context factory.
-     *
-     * @param className can be any of the qpid/amq supported JNDI properties:
-     *                  connectionfactory, queue, topic, destination.
-     * @param address   of the connection or node to create.
-     */
-    private Object createJMSProviderObject(String className, String address) {
-    /* Name of the object is the same as class of the object */
-        String name = className;
-        properties.setProperty(className + "." + name, address);
-
-        Object jmsProviderObject = null;
-        try {
-            Context context = new InitialContext(properties);
-            jmsProviderObject = context.lookup(name);
-            context.close();
-        } catch (NamingException e) {
-            e.printStackTrace();
-        }
-        return jmsProviderObject;
-    }
-
-    protected void setInitialContext(String userConnectionFactory) {
-        properties = new Properties();
-        try {
-            String jndiFilePath;
-            if ((jndiFilePath = System.getProperty(EXTERNAL_JNDI_PROPERTY)) != null) {
-                // load property file from an absolute path to the file
-                try (FileInputStream stream = new FileInputStream(jndiFilePath)) {
-                    properties.load(stream);
-                }
-            } else {
-                ClassLoader classloader = Thread.currentThread().getContextClassLoader();
-                String JNDIFile;
-                if (CoreClient.isAMQClient()) {
-                    JNDIFile = "QpidJmsJNDI.properties";
-                } else if (CoreClient.isQpidClient()) {
-                    JNDIFile = "QpidLegacyJNDI.properties";
-                } else {
-                    JNDIFile = "JAMQ6JNDI.properties";
-                }
-                LOG.trace("Using " + JNDIFile);
-                // fallback to use resources/ArtemisJNDI.properties file
-                properties.load(classloader.getResourceAsStream(JNDIFile));
-                LOG.trace(properties.toString());
-            }
-            LOG.trace(userConnectionFactory);
-            if (userConnectionFactory.contains("://")) {
-                // override connectionFactory by this option in jndi/properties
-                properties.setProperty(Context.PROVIDER_URL, userConnectionFactory);
-                properties.setProperty("connectionFactory." + amqConnectionFactoryJNDI, userConnectionFactory);
-            }
-            initialContext = new InitialContext(properties);
-        } catch (NamingException | IOException e) {
-            e.printStackTrace();
         }
     }
 }
